refactor: eliminate unnecessary as-casts via shared type inference
Add src/inference.js with inferExprType — a shared utility that resolves
return types from literals, typed identifiers, binary expressions (with
numeric promotion), built-in Text/Blob/GraphemeText methods, and SizeExpr.
Integrated into all three codegen backends (JS, Erlang, Rust) and the
extract service so that reply fields no longer require explicit as-casts
when the type is locally inferrable.

Removed ~1,100 unnecessary as-casts across 70 test files. Remaining casts
are legitimately needed: lambda return types for Rust codegen, as-feature
tests, spread/untyped destructure contexts, and remote method call results.

+++ __tests__/constructors/dependency_injection.test.js
@@ -187,7 +187,7 @@ describe('explicit form — full roundtrip', () => {
         :n Integer
         =
         :result Integer = m.double(:n)
-        -> answer: (result + 1) as Integer
+        -> answer: result + 1
     `, {
       expects: [
         { output: expect.objectContaining({ op: [{ base: 0 }, 'new'], to: 'Math' }) },
@@ -322,7 +322,7 @@ describe('# form — requires manifest', () => {
     expect(() => compileSource(`
       < "thing.bv": (Thing) # >
 
-      @go = -> 1 as Integer
+      @go = -> 1
     `)).toThrow(/requires an interface/);
   });
